requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.10.5
django-cors-headers==2.0.2
django-filter==1.0.1
django-rest-multiple-models==1.8.2
django-solo==1.1.2
djangorestframework==3.5.3
enum34==1.1.6
graphene==1.1.3
graphene-django==1.2.1
graphql-core==1.0.1
graphql-relay==0.4.5
gunicorn==19.6.0
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
iso8601==0.1.11
olefile==0.44
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
promise==1.0.1
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
redis==2.10.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
traitlets==4.3.1
typing==3.5.3.0
wcwidth==0.1.7
